chore(tests): split tests to avoid flakiness
